+++ CHANGELOG.rst
@@ -2,6 +2,25 @@
 CHANGELOG
 =========
 
+Next Release (TBD)
+==================
+
+* Add experimental support for websockets
+  (`#1017 <https://github.com/aws/chalice/issues/1017>`__)
+
+* API Gateway Endpoint Type Configuration
+  (`#1160 https://github.com/aws/chalice/pull/1160`__)
+
+* API Gateway Resource Policy Configuration
+  (`#1160 https://github.com/aws/chalice/pull/1160`__)
+
+1.9.1
+=====
+
+* Make MultiDict mutable
+  (`#1158 <https://github.com/aws/chalice/issues/1158>`__)
+
+
 1.9.0
 =====